requirements.txt
Django==2.0.3
psycopg2
django-bootstrap4
pytz
pdfkit
stura_voting_utils
schulze_voting
median_voting
python-dateutil
